requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
tenacity>=8.2.0
schedule>=1.2.0
python-telegram-bot>=20.0
yfinance>=0.2.0
newsapi-python>=0.2.7
selenium>=4.15.0
webdriver-manager>=4.0.0
